Wg 270 uc 56 view workflow

+++ backend/src/controllers/stage.controller.js
@@ -19,11 +19,13 @@ const getStagesByProjectId = async (req, res, next) => {
 const updateStageParents = async (req, res, next) => {
     try {
         const { updates } = req.body;
+        const { projectId } = req.params;
+        const userId = req.payload.id;
         if (!Array.isArray(updates) || updates.length === 0) {
             return res.status(400).json({ message: "Invalid request data." });
         }
 
-        const result = await stageService.updateStageParents(updates);
+        const result = await stageService.updateStageParents(updates, userId, projectId);
         res.status(200).json({ status: 200, ...result });
     } catch (error) {
         next(error);
@@ -34,12 +36,13 @@ const addStage = async (req, res, next) => {
     try {
         const { stageName, stageStatus, parent } = req.body;
         const { projectId } = req.params;
+        const userId = req.payload.id;
 
         if (!stageName) {
             return res.status(400).json({ message: "Stage name is required" });
         }
 
-        const newStage = await stageService.addStage({ stageName, stageStatus, parent, projectId });
+        const newStage = await stageService.addStage({ stageName, stageStatus, parent, projectId , userId});
 
         res.status(201).json({ status: 201, message: "Stage created successfully", stage: newStage });
     } catch (error) {
@@ -51,7 +54,7 @@ const addStage = async (req, res, next) => {
 const updateStage = async (req, res, next) => {
     try {
         const { stageId, stageName, stageStatus } = req.body; 
-
+        const userId = req.payload.id;
         if (!stageId) {
             return res.status(400).json({ error: { status: 400, message: "Stage ID is required" } });
         }
@@ -66,7 +69,7 @@ const updateStage = async (req, res, next) => {
             return res.status(400).json({ error: { status: 400, message: "Stage name cannot be empty" } });
         }
 
-        const updatedStage = await stageService.editStage(stageId, { stageName, stageStatus });
+        const updatedStage = await stageService.editStage(stageId, { stageName, stageStatus }, userId);
 
         if (!updatedStage) {
             return res.status(400).json({ error: { status: 400, message: "Stage update failed" } });
@@ -86,12 +89,13 @@ const deleteStage = async (req, res) => {
     try {
         const { stageId, targetStageId } = req.body;
         const { projectId } = req.params;
+        const userId = req.payload.id;
 
         if (!stageId || !projectId) {
             return res.status(400).json({ message: "Stage ID and Project ID are required." });
         }
 
-        const result = await stageService.deleteStage({ stageId, projectId, targetStageId });
+        const result = await stageService.deleteStage({ stageId, projectId, targetStageId ,userId});
 
         res.status(200).json({ status: 200, ...result });
     } catch (error) {
